move common functionality in etw_config.h
