+++ exporters/etw/include/opentelemetry/exporters/etw/etw_logger_exporter.h
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 #pragma once
 
+#ifdef ENABLE_LOGS_PREVIEW
 #include <cstdint>
 #include <cstdio>
 #include <cstdlib>
@@ -13,18 +14,20 @@
 #include "opentelemetry/nostd/unique_ptr.h"
 
 #include "opentelemetry/common/key_value_iterable_view.h"
-#include "opentelemetry/trace/span.h"
-#include "opentelemetry/trace/span_context_kv_iterable_view.h"
+
 #include "opentelemetry/trace/span_id.h"
 #include "opentelemetry/trace/trace_id.h"
-#include "opentelemetry/trace/tracer_provider.h"
+#include "opentelemetry/logs/tracer_provider.h"
 
-#include "opentelemetry/sdk/trace/exporter.h"
+#include "opentelemetry/sdk/logs/exporter.h"
 
 #include "opentelemetry/exporters/etw/etw_provider.h"
 #include "opentelemetry/exporters/etw/etw_config.h"
-#include "opentelemetry/exporters/etw/etw_tracer.h"
+#include "opentelemetry/exporters/etw/etw_logger.h"
 
 #include "opentelemetry/exporters/etw/utils.h"
 
+namespace logs = opentelemetry::logs;
 namespace trace = opentelemetry::trace;
+#endif // ENABLE_LOGS_PREVIEW
+
