+++ exporters/etw/test/etw_perf_test.cc
@@ -20,7 +20,7 @@ namespace
 {
 static constexpr const char *providerName = "OpenTelemetry-ETW-StressTest";
 
-static exporter::etw::TracerProviderOptions providerOptions = {{"enableTraceId", false},
+static exporter::etw::TelemetryProviderOptions providerOptions = {{"enableTraceId", false},
                                                                {"enableSpanId", false},
                                                                {"enableActivityId", false},
                                                                {"enableRelatedActivityId", false},
